chore: rename /golang/cosmos/chain to vm
