+++ golang/cosmos/vm/controller.go
@@ -1,4 +1,4 @@
-package chain
+package vm
 
 import (
 	"errors"
