+++ golang/cosmos/vm/height.go
@@ -1,4 +1,4 @@
-package chain
+package vm
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
