+++ golang/cosmos/x/dibc/handler.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/Agoric/agoric-sdk/golang/cosmos/chain"
+	"github.com/Agoric/agoric-sdk/golang/cosmos/vm"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
@@ -12,7 +12,7 @@ import (
 // NewHandler returns a handler for "dibc" type messages.
 func NewHandler(keeper Keeper) sdk.Handler {
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		if chain.IsSimulation(ctx) {
+		if vm.IsSimulation(ctx) {
 			// We don't support simulation.
 			return &sdk.Result{}, nil
 		} else {
