+++ golang/cosmos/x/vpurse/handler.go
@@ -3,7 +3,7 @@ package vpurse
 import (
 	"fmt"
 
-	"github.com/Agoric/agoric-sdk/golang/cosmos/chain"
+	"github.com/Agoric/agoric-sdk/golang/cosmos/vm"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
@@ -11,7 +11,7 @@ import (
 // NewHandler returns a handler for "vpurse" type messages.
 func NewHandler(keeper Keeper) sdk.Handler {
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		if chain.IsSimulation(ctx) {
+		if vm.IsSimulation(ctx) {
 			// We don't support simulation.
 			return &sdk.Result{}, nil
 		} else {
